tests/moulinette/requirements.txt
Click
future==0.18.2
argparse
dataclasses
PyYAML
pytest
termcolor
